Completed day 5 of 2023 and fixed a warning in day 4

+++ source/challenges/2023/Challenge04.h
@@ -112,7 +112,7 @@ namespace Year2023
             }
             else
             {
-                return static_cast<int>(powf(2.f, card.m_matchCount - 1));
+                return static_cast<int>(powf(2.f, static_cast<float>(card.m_matchCount - 1)));
             }
         }
     };
